ya tenemos el api funcionando  solo falta meter la info en el html / handle bars
