+++ views/layouts/main.handlebars
@@ -18,9 +18,10 @@
 
   {{> footer}}
  
-        {{#each scripts}}
-            <script src="{{this}}"></script>
-        {{/each}}
+       
 
 </body>
+ {{#each scripts}}
+            <script src="{{this}}"></script>
+        {{/each}}
 </html>
